// src/payload/payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Gallery {
  id: number;
  title?: string | null;
  description?: string | null;
  category?: ('sports' | 'arts' | 'academic' | 'campus' | 'events') | null;
  image?: (number | null) | Media;
  date?: string | null;
  featured?: boolean | null;
  updatedAt: string;
  createdAt: string;
}